# web/src/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Ea.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/systems/{id}/integrations": {
      "get": {
        "tags": [
          "Integrationer"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemIntegrationsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/integrations/export.csv": {
      "get": {
        "tags": [
          "Integrationer"
        ],
        "parameters": [
          {
            "name": "systemId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/integrations/{id}": {
      "get": {
        "tags": [
          "Integrationer"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Integrationer"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Integrationer"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/integrations": {
      "post": {
        "tags": [
          "Integrationer"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataObjectDto": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "IntegrationCreateRequest": {
        "required": [
          "fromSystemId",
          "toSystemId",
          "viaPlatformId",
          "type",
          "name",
          "description",
          "dataObjectIds"
        ],
        "type": "object",
        "properties": {
          "fromSystemId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "toSystemId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "viaPlatformId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IntegrationType"
              }
            ]
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "dataObjectIds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "IntegrationDto": {
        "required": [
          "id",
          "name",
          "from",
          "to",
          "via",
          "type",
          "description",
          "dataObjects",
          "createdAt",
          "updatedAt",
          "version",
          "permissions"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "from": {
            "$ref": "#/components/schemas/SystemLink"
          },
          "to": {
            "$ref": "#/components/schemas/SystemLink"
          },
          "via": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SystemLink"
              }
            ]
          },
          "type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IntegrationType"
              }
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "dataObjects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataObjectDto"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "format": "uint32"
          },
          "permissions": {
            "$ref": "#/components/schemas/IntegrationPermissions"
          }
        }
      },
      "IntegrationPermissions": {
        "required": [
          "canEdit"
        ],
        "type": "object",
        "properties": {
          "canEdit": {
            "type": "boolean"
          }
        }
      },
      "IntegrationRelation": {
        "enum": [
          "Ud",
          "Via",
          "Ind",
          "Intern"
        ]
      },
      "IntegrationSummary": {
        "required": [
          "receivers",
          "suppliers",
          "viaPlatform",
          "localSolutions",
          "directDb"
        ],
        "type": "object",
        "properties": {
          "receivers": {
            "type": "integer",
            "format": "int32"
          },
          "suppliers": {
            "type": "integer",
            "format": "int32"
          },
          "viaPlatform": {
            "type": "integer",
            "format": "int32"
          },
          "localSolutions": {
            "type": "integer",
            "format": "int32"
          },
          "directDb": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "IntegrationType": {
        "enum": [
          "Api",
          "Fil",
          "Event",
          "DirekteDb",
          "Udtraek",
          null
        ]
      },
      "IntegrationUpdateRequest": {
        "required": [
          "viaPlatformId",
          "type",
          "name",
          "description",
          "dataObjectIds",
          "version"
        ],
        "type": "object",
        "properties": {
          "viaPlatformId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IntegrationType"
              }
            ]
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "dataObjectIds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "version": {
            "type": [
              "null",
              "integer"
            ],
            "format": "uint32"
          }
        }
      },
      "LifecycleStatus": {
        "enum": [
          "Planlagt",
          "Indfases",
          "IDrift",
          "Udfases",
          "Nedlagt"
        ]
      },
...
      "SystemIntegrationItem": {
        "required": [
          "relation",
          "counterpart",
          "localModule",
          "integration"
        ],
        "type": "object",
        "properties": {
          "relation": {
            "$ref": "#/components/schemas/IntegrationRelation"
          },
          "counterpart": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SystemLink"
              }
            ]
          },
          "localModule": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SystemRef"
              }
            ]
          },
          "integration": {
            "$ref": "#/components/schemas/IntegrationDto"
          }
        }
      },
      "SystemIntegrationsResponse": {
        "required": [
          "summary",
          "items",
          "canAdd"
        ],
        "type": "object",
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/IntegrationSummary"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SystemIntegrationItem"
            }
          },
          "canAdd": {
            "type": "boolean"
          }
        }
      },
      "SystemLink": {
        "required": [
          "id",
          "name",
          "parent",
          "type",
          "lifecycleStatus"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "parent": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SystemRef"
              }
            ]
          },
          "type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SystemType"
              }
            ]
          },
          "lifecycleStatus": {
            "$ref": "#/components/schemas/LifecycleStatus"
          }
        }
      },
...
      "SystemRef": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "SystemType": {
        "enum": [
          "Saas",
          "Standardsystem",
          "Egenudviklet",
          "LowCode",
          "Platform",
          "LokalLoesning",
          null
        ]
      },
...
    }
  },
...
}